# docs/sms-api-doc.yaml
openapi: 3.0.1
info:
  title: School Management System API
  version: "1.3.3"
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
...
  /parents:
    get:
      tags:
      - Parents
      summary: Find all parents
      description: "Retrieve all existing parents in the system. Required authority\
        \ level: Administrator"
      operationId: getParents
      responses:
        "200":
          description: Returns existing objects list (empty result as well)
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ParentSummaryResponse'
        "401":
          description: Full authentication is required to access this resource
        "403":
          description: User lacks permission to access this resource
      security:
      - bearerAuth: []
    post:
      tags:
      - Parents
      summary: Create parent
      description: "Adds new parent profile related with children. Response includes\
        \ first generated password. Required authority level: Administrator"
      operationId: addParent
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ParentCreateRequest'
        required: true
      responses:
        "201":
          description: Created successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AddParentResponse'
        "401":
          description: Full authentication is required to access this resource
        "404":
          description: Children not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AddParentResponse'
        "403":
          description: User lacks permission to access this resource
        "406":
          description: Authority with given ID not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AddParentResponse'
        "409":
          description: User with given email already exists
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AddParentResponse'
      security:
      - bearerAuth: []
...
  /schedules/{scheduleId}:
    delete:
      tags:
      - Schedules
      summary: Delete schedule
      description: "Remove existing schedule if has no ongoing occurrences, and cancels\
        \ booked occurrences. Required authority level: Administrator"
      operationId: deleteSchedule
      parameters:
      - name: scheduleId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "406":
          description: Active occurrences ongoing
          content:
            '*/*':
              schema:
                type: string
        "204":
          description: Removed successfully
          content:
            '*/*':
              schema:
                type: string
        "401":
          description: Full authentication is required to access this resource
        "403":
          description: User lacks permission to access this resource
        "404":
          description: Not found by ID
          content:
            '*/*':
              schema:
                type: string
      security:
      - bearerAuth: []
    patch:
      tags:
      - Schedules
      summary: Update schedule
      description: "Modify existing schedule, if time period is applicable and already\
        \ booked. Required authority level: Administrator"
      operationId: patchSchedule
      parameters:
      - name: scheduleId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchScheduleRequest'
        required: true
      responses:
        "406":
          description: Modification canceled because of actual schedule status
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ScheduleDetailedResponse'
        "200":
          description: Modification successful
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ScheduleDetailedResponse'
        "401":
          description: Full authentication is required to access this resource
        "404":
          description: Not found by ID
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ScheduleDetailedResponse'
        "403":
          description: User lacks permission to access this resource
      security:
      - bearerAuth: []
  /parents/{parentId}:
    get:
      tags:
      - Parents
      summary: Find parent with ID
      description: "Retrieves existing parent with given ID. Required authority level:\
        \ Administrator, Teacher, Student, Parent"
      operationId: getParent
      parameters:
      - name: parentId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "404":
          description: Not found with ID
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ParentDetailedResponse'
        "200":
          description: Returns existing object
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ParentDetailedResponse'
        "401":
          description: Full authentication is required to access this resource
        "403":
          description: User lacks permission to access this resource
      security:
      - bearerAuth: []
    delete:
      tags:
      - Parents
      summary: Delete parent
      description: "Remove parent with given ID and unlink children. Required authority\
        \ level: Administrator"
      operationId: deleteParent
      parameters:
      - name: parentId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "204":
          description: Removed successfully
          content:
            '*/*':
              schema:
                type: string
        "404":
          description: Not found with ID
          content:
            '*/*':
              schema:
                type: string
        "401":
          description: Full authentication is required to access this resource
        "403":
          description: User lacks permission to access this resource
      security:
      - bearerAuth: []
    patch:
      tags:
      - Parents
      summary: Update parent
      description: "Modify parent profile and relation with children. Required authority\
        \ level: Administrator"
      operationId: updateParent
      parameters:
      - name: parentId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ParentPatchRequest'
        required: true
      responses:
        "404":
          description: Parent not found with ID
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ParentDetailedResponse'
        "200":
          description: Modification successfull
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ParentDetailedResponse'
        "406":
          description: Child association removal - child not linked with parent
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ParentDetailedResponse'
        "401":
          description: Full authentication is required to access this resource
        "403":
          description: User lacks permission to access this resource
      security:
      - bearerAuth: []
...
  /schedules/{scheduleVersionId}/{scheduleId}:
    get:
      tags:
      - Schedules
      summary: Find schedule by ID
      description: "Find existing schedule with ID. Required authority level: Administrator,\
        \ Teacher, Student"
      operationId: getSchedule
      parameters:
      - name: scheduleId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      - name: scheduleVersionId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Returns existing object
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ScheduleDetailedResponse'
        "401":
          description: Full authentication is required to access this resource
        "403":
          description: User lacks permission to access this resource
        "404":
          description: Not found by ID
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ScheduleDetailedResponse'
      security:
      - bearerAuth: []
...
  /occurrences/{scheduleOccurrenceId}:
    get:
      tags:
      - Schedule Occurrences
      summary: Get occurrence with ID
      description: "Retrieves existing schedule occurrence with given ID. Required\
        \ authority level: Administrator"
      operationId: getOccurrenceById
      parameters:
      - name: scheduleOccurrenceId
        in: query
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not found with given ID
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ScheduleOccurrenceSummaryResponse'
        "200":
          description: Returns existing object
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ScheduleOccurrenceSummaryResponse'
        "401":
          description: Full authentication is required to access this resource
        "403":
          description: User lacks permission to access this resource
      security:
      - bearerAuth: []
  /occurrences/yearbook/{yearbookId}:
    get:
      tags:
      - Schedule Occurrences
      summary: Get schedule occurrences for yearbook
      description: "Retrieves all schedule (active) occurrences planned for given\
        \ Yearbook in particular time range. Time range is configured in app configuration\
        \ value. Required authority level: Administrator, Teacher, Student, Parent"
      operationId: getOccurrencesForYearbook
      parameters:
      - name: yearbookId
        in: query
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "404":
          description: Required app config not found
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ScheduleOccurrenceSummaryResponse'
        "401":
          description: Full authentication is required to access this resource
        "406":
          description: Cannot fetch occurrences because there is no active schedule
            version
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ScheduleOccurrenceSummaryResponse'
        "403":
          description: User lacks permission to access this resource
        "200":
          description: Returns objects list (empty result as well)
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ScheduleOccurrenceSummaryResponse'
      security:
      - bearerAuth: []
...
components:
  schemas:
...
    StudentSummaryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int32
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        birthDate:
          type: string
          format: date-time
        gender:
          type: string
          enum:
          - MALE
          - FEMALE
        phoneNumber:
          type: string
    SubjectSummaryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        code:
          type: string
        description:
          type: string
        subjectType:
          type: string
    TeacherSummaryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int32
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        birthDate:
          type: string
          format: date-time
        gender:
          type: string
          enum:
          - MALE
          - FEMALE
        phoneNumber:
          type: string
...
    AddressRequest:
      type: object
      properties:
        city:
          type: string
        street:
          type: string
        buildingNumber:
          type: string
        voivodeship:
          type: string
        zipCode:
          type: string
    AuthorityEntity:
      type: object
      properties:
        id:
          type: integer
          format: int32
        authority:
          type: string
...
    AddressResponse:
      type: object
      properties:
        id:
          type: integer
          format: int32
        city:
          type: string
        street:
          type: string
        buildingNumber:
          type: string
        voivodeship:
          type: string
        zipCode:
          type: string
...
    LocalTime:
      type: object
      properties:
        hour:
          type: integer
          format: int32
        minute:
          type: integer
          format: int32
        second:
          type: integer
          format: int32
        nano:
          type: integer
          format: int32
...
    ParentSummaryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int32
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        birthDate:
          type: string
          format: date-time
        gender:
          type: string
          enum:
          - MALE
          - FEMALE
        phoneNumber:
          type: string
...
    ClassroomSummaryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        capacity:
          type: integer
          format: int32
        equipmentCount:
          type: integer
          format: int32
    ScheduleDetailedResponse:
      type: object
      properties:
        id:
          type: integer
          format: int32
        subject:
          $ref: '#/components/schemas/SubjectSummaryResponse'
        teacher:
          $ref: '#/components/schemas/TeacherSummaryResponse'
        classroom:
          $ref: '#/components/schemas/ClassroomSummaryResponse'
        dayOfWeek:
          type: string
          enum:
          - MONDAY
          - TUESDAY
          - WEDNESDAY
          - THURSDAY
          - FRIDAY
          - SATURDAY
          - SUNDAY
        startTime:
          $ref: '#/components/schemas/LocalTime'
        endTime:
          $ref: '#/components/schemas/LocalTime'
        recurrenceType:
          type: string
          enum:
          - NONE
          - WEEKLY
          - BIWEEKLY
          - MONTHLY
        scheduleVersion:
          $ref: '#/components/schemas/ScheduleVersionSummaryResponse'
        status:
          type: string
          enum:
          - SCHEDULED
          - CANCELLED
          - RESCHEDULED
          - DELETED
          - UPDATED
        effectiveDate:
          type: string
          format: date
        expirationDate:
          type: string
          format: date
...
    ScheduleVersionSummaryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        active:
          type: boolean
...
    ScheduleOccurrenceSummaryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        schedule:
          $ref: '#/components/schemas/ScheduleSummaryResponse'
        occurrenceDateTime:
          type: string
          format: date-time
        occurrenceEndTime:
          $ref: '#/components/schemas/LocalTime'
        status:
          type: string
          enum:
          - SCHEDULED
          - ONGOING
          - COMPLETED
          - CANCELLED
    ScheduleSummaryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int32
        dayOfWeek:
          type: string
          enum:
          - MONDAY
          - TUESDAY
          - WEDNESDAY
          - THURSDAY
          - FRIDAY
          - SATURDAY
          - SUNDAY
        startTime:
          $ref: '#/components/schemas/LocalTime'
        endTime:
          $ref: '#/components/schemas/LocalTime'
        classroomName:
          type: string
        teacherName:
          type: string
        subjectName:
          type: string
        scheduleVersionId:
          type: integer
          format: int32
        status:
          type: string
          enum:
          - SCHEDULED
          - CANCELLED
          - RESCHEDULED
          - DELETED
          - UPDATED
    ParentCreateRequest:
      required:
      - address
      - authority
      - birthDate
      - childrenIds
      - email
      - firstName
      - lastName
      type: object
      properties:
        email:
          type: string
        phoneNumber:
          type: string
        firstName:
          maxLength: 50
          minLength: 1
          type: string
        lastName:
          maxLength: 50
          minLength: 1
          type: string
        birthDate:
          type: string
          format: date-time
        gender:
          type: string
          enum:
          - MALE
          - FEMALE
        address:
          $ref: '#/components/schemas/AddressRequest'
        authority:
          type: array
          items:
            $ref: '#/components/schemas/AuthorityEntity'
        dateOfCreation:
          type: string
          format: date-time
        childrenIds:
          type: array
          items:
            type: integer
            format: int32
    AddParentResponse:
      type: object
      properties:
        id:
          type: integer
          format: int32
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        phoneNumber:
          type: string
        birthDate:
          type: string
          format: date-time
        gender:
          type: string
          enum:
          - MALE
          - FEMALE
        authority:
          type: array
          items:
            $ref: '#/components/schemas/AuthorityEntity'
        address:
          $ref: '#/components/schemas/AddressResponse'
        children:
          type: array
          items:
            $ref: '#/components/schemas/StudentSummaryResponse'
        password:
          type: string
...
    PatchScheduleRequest:
      type: object
      properties:
        subjectId:
          type: integer
          format: int32
        teacherId:
          type: integer
          format: int32
        classroomId:
          type: integer
          format: int32
        yearbookId:
          type: integer
          format: int32
        dayOfWeek:
          maximum: 7
          minimum: 1
          type: integer
          format: int32
        startTime:
          $ref: '#/components/schemas/LocalTime'
        endTime:
          $ref: '#/components/schemas/LocalTime'
        recurrenceType:
          type: string
          enum:
          - NONE
          - WEEKLY
          - BIWEEKLY
          - MONTHLY
        updateReason:
          type: string
        effectiveDate:
          type: string
          format: date
        expirationDate:
          type: string
          format: date
        validExpirationDate:
          type: boolean
        validEffectiveDate:
          type: boolean
        effectiveDateBeforeExpirationDate:
          type: boolean
...
    ParentPatchRequest:
      type: object
      properties:
        email:
          type: string
        phoneNumber:
          type: string
        firstName:
          maxLength: 50
          minLength: 1
          type: string
        lastName:
          maxLength: 50
          minLength: 1
          type: string
        birthDate:
          type: string
          format: date-time
        gender:
          type: string
          enum:
          - MALE
          - FEMALE
        address:
          $ref: '#/components/schemas/AddressRequest'
        authority:
          type: array
          items:
            $ref: '#/components/schemas/AuthorityEntity'
        dateOfUpdate:
          type: string
          format: date-time
        childrenIdsToAdd:
          type: array
          items:
            type: integer
            format: int32
        childrenIdsToRemove:
          type: array
          items:
            type: integer
            format: int32
    ParentDetailedResponse:
      type: object
      properties:
        id:
          type: integer
          format: int32
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        phoneNumber:
          type: string
        birthDate:
          type: string
          format: date-time
        gender:
          type: string
          enum:
          - MALE
          - FEMALE
        authority:
          type: array
          items:
            $ref: '#/components/schemas/AuthorityEntity'
        address:
          $ref: '#/components/schemas/AddressResponse'
        children:
          type: array
          items:
            $ref: '#/components/schemas/StudentSummaryResponse'
...
  securitySchemes:
    bearerAuth:
      type: http
      name: bearerAuth
      scheme: bearer
      bearerFormat: JWT